requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pycryptodome==3.15.0
pyOpenSSL==22.1.0
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
